Use Prelude.String's in the Sing (t :: JType) instance.
Fixes #11.

The singleton instance used JNI.String, which is a wrapper over
ByteStrings, relies on unsafePerformIO and related functions to do
comparisons. For some reason, these calls are not always floated to
the top-level, which means that they are redone everytime a
comparison needs to be performed.

Redoing comparisons hurts performance of call and callStatic in
the jvm package.

When using Prelude.String instead, the comparisons become pure,
which allows their results to be floated and cached when call
and callStatic are evaluated multiple times at the same types.