Make the result of JNI.String constructors bear the proof of well-formation.
678d5b1416df435f84fe31ff1110eb15480f96c6 showed how it is all too easy
to pass an ill-formed class name to findClass. We do get a runtime
error in this case but sometimes the error returned by the JVM is
confusing. Better to statically rule out the possibility of
a well-formation issue to begin with. That's what the abstract newtype
wrappers introduced here seek to achieve: extra safety and piece of
mind.

This is a backward compatibility breaking change.