Pass class name as a JNI.String to callStatic, not as a singleton.
The original rationale for passing in a singleton was that like `call`
and other functions, it becomes possible to `SPECIALIZE` `callStatic`
to certain classes. But this isn't terribly useful, indeed entirely
redundant with what could be achieved with simple inlinining, hence
making the added syntactic overhead unjustifiable.

```
callStatic (sing :: Sing "some.java.Class") ...
```

is now

```
callStatic "some.java.Class" ...
```

This is a breaking change. Best to change this now before this really
becomes enshrined.