Provide defaults for reify/reflect.
A typical pattern is to wrap Java objects in newtypes on the Haskell
side to expose them as first-class abstract Haskell values. These
objects can be unwrapped easily enough using the `Coercible` class or
otherwise. But whereas Java objects are trivially
reifiable/reflectable, their newtype wrappers aren't. With this
change, we make it easier to derive `Reify`/`Reflect` instances for
these newtype wrappers. Having these instances defined is necessary to
uniformly pass any value to functions that want to reflect it,
including when the value is itself a newtype wrapper around a java
object.

See sparkle#93 for a good example use case that would be made more
concise thanks to this change.