Do away with ghc-heap-view dependency.
The static pointer table (SPT) in GHC currently lacks any typing
information. So to discriminate between static pointers that refer to
inline-java bytecode vs other random static pointers, we used
ghc-heap-view as a workaround.

It was a gross hack. It worked. But the cost of using this hammer is
that profiling builds were impossible.

This commit instead uses a much smaller hammer. Where ghc-heap-view
requires very intimate knowledge of much of GHC's memory layout, we
only use 'unpackClosure#'. Because we don't have the full power of
ghc-heap-view, we can't inspect the info table provided to us by this
primitive. But what we can do is inspect the environment of the
closure. If we find a pair of magic integers in there, then we assume
the static pointer does point to Java bytecode.