Fix NUL handling in JNI strings.
There are two issues addressed by this patch.

The first problem is that fromChars, yielded a bytestring which
didn't include the terminating NUL character. The bytestring,
however, pointed to a C string that was NUL terminated, and this
made the representation to work most of the time.

toByteString, in addition, was not discarding the terminating
NUL character, thus

toByteString (fromChars a) `BS.append` ";"

would contain a NUL character before ";" when fromChars is fixed.

The second problem that is addressed, is that formerly, the memory
used by JNI strings lived in the C heap, putting no pressure on the
Haskell GC to collect them. Now they live in the Haskell heap.