Rewrite environment checking.
It's a miracle to me that the previous code worked somehow. But
I believe this version is clearer. The key invariant is spelled out
explicitly, and `getJNIEnv` no longer has myriad different cases to
check for, thanks to the rest of the code being engineered to respect
this invariant. `newJVM` now stores the JNIEnv in a thread-local
variable, just like `attach*` does.