Use standard directory layout used by Maven for multi-language packages.
We put Haskell source in src/main/haskell and Java source in
src/main/java. The point of doing this is that we can further separate
out Cabal specific source files in src/cabal/haskell. It's useful to
separate these out because they use a Cabal-specific API, which is not
supported by rules_haskell.

See https://maven.apache.org/guides/introduction/introduction-to-the-standard-directory-layout.html