Rename all BUILD files to BUILD.bazel.
This is a workaround for an unfortunate situation. On macOS, the
filesystem is insensitive by default. So `BUILD` and `build` are the
same thing. Now, it so happens that `build` is also the name of the
directory that Gradle wants to put build artifacts in. Gradle fails on
macOS (but not Linux) because it can't create a `build` folder,
because `build` is already a file, i.e. a Bazel `BUILD` file.