Workaround Gradle sometimes trashing output in recent versions.
Newer versions of Gradle think it's a great idea to output a summary
of the release notes even when the `-q` flag for "quiet!" is passed,
but only on the first run. This trashes the output that we're
capturing to get at the CLASSPATH. So as a workaround, we make sure to
only capture the *last* line of output from `gradle build -q`.