jvm: make call* variadic
`call` previously took a list of `JValue`'s. Using the dictionaries
embedded in each element, it could construct a `MethodID`. However, this
`MethodID` was not cached as well as one would hope, say when multiple
calls are made to the same method at the same argument types but
different values. Another problem was that all arguments to the method
had to be packed into a list and coerced into a `JValue` before being
passed to the method. This was syntactic overhead. The solution is to
make `call`, `new` and `callStatic` variadic, using the same GADT
solution as for variadic `printf`.

We now pass an extra spec argument to `call`. Like the format string in
`printf`, this argument determines how many more arguments need to be
passed to `call`. Example:

```haskell
call obj "frobnicate" With3Args x y z
```

We no longer need to coerce `x`, `y`, `z`, nor pack them into a list.
Furthermore, this solution leads to decent error messages if arguments
are missing.