jvm: make call* variadic
`call` previously took a list of `JValue`'s. Using the dictionaries
embedded in each element, it could construct a `MethodID`. However, this
`MethodID` was not cached as well as one would hope, say when multiple
calls are made to the same method at the same argument types but
different values. Another problem was that all arguments to the method
had to be packed into a list and coerced into a `JValue` before being
passed to the method. This was syntactic overhead. The solution is to
make `call`, `new` and `callStatic` variadic, using the same GADT
solution as for variadic `printf`.

Using ad hoc private type classes per function, we make them variadic up
to 32 arguments.

```haskell
call obj "frobnicate" x y z
```

We no longer need to coerce `x`, `y`, `z`, nor pack them into a list.
Furthermore, this solution leads to decent error messages. The private
type class should never appear in error messages. It's hard to think of
any example where the error message would be something else than "could
not deduce `Coercible a` ...". Which is good enough, since it's telling
us that you can pass anything you want to `call`, but it must be
coercible.