Move Stack and Bazel linux jobs to buildkite
Getting the cache key to be adequately conservative in CircleCI is a
nightmare. And even with caching, reinstalling all dependencies +
restorting the cache takes so much time that an identity build takes 10
minutes or more. The solution is to use BuildKite, like rules_haskell
and rules_sh already do. This won't work (yet) for macOS. But for the
Linux jobs this should speed things up quite significantly, while also
making the CI job definitions simpler.