Introduce Variadic constraint
Variadic functions all have a constraint, which appears in their type
signature. In an effort to further hide this constraint, we replace it
in each case with a generic `Variadic "foo" r` constraint, where `foo`
is the name of the variadic function. That way we can explain in
a single place what this one constraint means throughout the module.

This constraint should only appear in Haddock. In principle, it should
never appear in error messages.