Introduce catchall instances for variadic functions error cases
Using TypeError, we can replace the error message that GHC provides to
the user when a constraint is not satisfied. We can in particular
arrange for such a custom message to be displayed whenever GHC would
instead talk about potentially private constraints that mean nothing
to the user.

This requires a catchall instance that overlaps with all others. The
overlap is benign, because the overlapping instances are for private
constraints that are never exported. Also, the error instance doesn't
have any operational meaning. It's just to throw an error at the user
when the program doesn't type check anyways.