jvm: Implement variadic arguments in the safe interface.
The design is complicated w.r.t. the unsafe interface because
the linear interface uses an abstract monad supplying an
instance of MonadIO.

There is firstly the challenge to pick the right instance when
types in the context are not fully instantiated. And secondly,
there are more situations where implementation details could
leak in error messages, and more custom error messages need
to be provided.

Closes #137.