Allow reifying batches which may contain null values.
A newtype Nullable is added to Language.Java.

"Vector (Nullable a)" is reified with batching. A batch
of "Nullable a" is a batch of "a"s together with a
boolean array that indicates whether the value at a
given position of the batch is null.

The reifyBatch method gets a new predicate as parameter, which
indicates whether a position in a batch should be reified.
Since composite values might be null, it doesn't make sense
to reify the corresponding positions in the batches of the
parts.