wip: introduce a GlobalRefCleaner data type
A GlobalRefCleaner holds an Async, and several synchronisations objects
to allow the deletion of global references by a dedicated thread,
attached to the JVM.

This choice was prefered to the previous implementation (a fork in the
jvm starting procedure) because the thread needs to be properly
terminated when the jvm stops.

Currently, the code does not compile due to typing issues.