fix: comment jvm-batching version number
Maven requires a version number for the artfacts, e.g. jvm-batching-0.2.0.jar
However, we reference the unversioned jar at several places in the codebase.
We would have to modify the name of the jar each time the version increases,
which would be tedious.

This version number would have to be uncommented before releasing to
Maven